consolidate normalizeAttachmentFilePath into signedUrlReplacer
Drop the standalone attachmentFilePath.ts and co-locate the helper with
getFilePathFromUrl in signedUrlReplacer.ts — both are URL/path normalization
helpers. Re-export from attachmentUtils.ts to preserve existing import paths.
